add handleErrorForm
